Allow installing with PHP-8 (#21)
* Allow installing with PHP-8
* Add workaround for PHPUnit dependencies
* Upgrade phpunit/phpunit
* Make PHPStan happy
* Make PHPUnit happy

+++ src/Base32/GmpEncoder.php
@@ -111,7 +111,7 @@ class GmpEncoder extends BaseEncoder
 
         return implode("", array_map(function ($byte) {
             //return pack("C", bindec($byte));
-            return chr(bindec($byte));
+            return chr((int)bindec($byte));
         }, $data));
     }
 
@@ -178,6 +178,6 @@ class GmpEncoder extends BaseEncoder
         }, $data);
         $binary = implode("", $data);
 
-        return bindec($binary);
+        return (int)bindec($binary);
     }
 }
